Revamp data processing pipeline
