+++ src/data_processing/augment_dataset.py
@@ -8,7 +8,7 @@ import zarr
 import numpy as np
 
 from src.models.vision import get_encoder
-from src.data_processing.process_demos import encode_demo
+from src.data_processing.process_pickles import encode_demo
 from src.visualization.render_mp4 import create_mp4
 
 base_dir = Path(os.environ["ROLLOUT_SAVE_DIR"])
