+++ src/models/mlp.py
@@ -1,4 +1,3 @@
-import torch
 import torch.nn as nn
